Working on removing factorial term
Made it so that the plugboard returns the coefficients for the power
series without the normalizing factorial term. However, there is
currently an issue where when we perform a linear combination of two
ODEs that we train the PINN on we cannot generate the same number of
coefficients as the training examples... very weird. But, the PINNs
datasets now have coefficients without the factorial term and we
multiply the coefficients at the end when we check the solution.